restructured pages
