+++ pages/plants/new.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PlantForm from '../../../components/PlantForm';
+import PlantForm from '../../components/PlantForm';
 
 export default function newPlant() {
   return (<PlantForm />);
