+++ pages/plants/viewAll.js
@@ -3,23 +3,23 @@
 import React, { useEffect, useState } from 'react';
 import Link from 'next/link';
 import { Button } from 'react-bootstrap';
-import { getAllPlants } from '../../../api/plantData';
-import PlantCard from '../../../components/PlantCard';
-import { useAuth } from '../../../utils/context/authContext';
+import { getAllPlants } from '../../api/plantData';
+import PlantCard from '../../components/PlantCard';
+// import { useAuth } from '../../../utils/context/authContext';
 
 export default function ViewAllPlants() {
 // Set state for plants
   const [plants, setPlants] = useState([]);
 
-  const { user } = useAuth();
+  // const { user } = useAuth();
 
   const getAllThePlants = () => {
-    getAllPlants(user.uid).then(setPlants);
+    getAllPlants().then(setPlants);
   };
 
   useEffect(() => {
     getAllThePlants();
-  }, [user.uid]);
+  }, []);
 
   return (
     <div className="text-center my-4">
