+++ pages/userProfile/new.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import UserProfileForm from '../../../components/UserProfileForm';
+import UserProfileForm from '../../components/UserProfileForm';
 
 export default function NewUserProfile() {
   return <UserProfileForm />;
