+++ utils/client.js
@@ -4,12 +4,6 @@ import 'firebase/auth';
 const firebaseCredentials = {
   apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
   authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.NEXT_PUBLIC_DATABASE_URL,
-
-  // projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-  // storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
-  // messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
-  // appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
 };
 
 const clientCredentials = {
